feat(cartes): footer sans sous-thème, pastille de couverture multi-sources
Le signal « ce sujet est couvert par N rédactions » était déjà calculé et servi
mais éteint par un gate trop étroit (isEssentiel && pressReviewCount > 0). La
puce est promue en CoverageChip partagé (avatars + « N sources »), gatée sur
sourceCount >= 2, et la pastille de sous-thème ML est retirée du footer.
