+++ apps/mobile/test/features/flux_continu/widgets/flux_continu_article_card_test.dart
@@ -5,6 +5,8 @@ import 'package:google_fonts/google_fonts.dart';
 
 import 'package:facteur/config/theme.dart';
 import 'package:facteur/features/feed/models/content_model.dart';
+import 'package:facteur/features/digest/models/digest_models.dart';
+import 'package:facteur/features/flux_continu/widgets/coverage_chip.dart';
 import 'package:facteur/features/flux_continu/widgets/flux_continu_article_card.dart';
 import 'package:facteur/features/settings/models/display_mode_spec.dart';
 import 'package:facteur/features/settings/providers/display_mode_provider.dart';
@@ -77,25 +79,63 @@ void main() {
     expect(find.byType(FacteurThumbnail), findsNothing);
   });
 
-  testWidgets(
-      'footer renders the theme pill from the source-theme fallback when '
-      'ML topics are empty (mitigation « tout en Autres »)', (tester) async {
+  testWidgets('footer shows the coverage chip when the subject is covered by '
+      'several newsrooms', (tester) async {
+    await tester.pumpWidget(_wrap(
+      FluxContinuArticleCard(
+        article: _content(),
+        sourceCount: 3,
+        perspectiveSources: const [
+          SourceMini(name: 'Le Monde'),
+          SourceMini(name: 'Libération'),
+          SourceMini(name: 'Le Figaro'),
+        ],
+      ),
+    ));
+
+    expect(find.byKey(const Key('flux-coverage-chip')), findsOneWidget);
+    expect(find.text('3 sources'), findsOneWidget);
+  });
+
+  testWidgets('no coverage chip below the 2-source threshold', (tester) async {
+    await tester.pumpWidget(_wrap(
+      FluxContinuArticleCard(
+        article: _content(),
+        sourceCount: 1,
+        perspectiveSources: const [SourceMini(name: 'Le Monde')],
+      ),
+    ));
+
+    expect(find.byKey(const Key('flux-coverage-chip')), findsNothing);
+  });
+
+  testWidgets('coverage chip caps the avatar stack at 3', (tester) async {
+    await tester.pumpWidget(_wrap(
+      FluxContinuArticleCard(
+        article: _content(),
+        sourceCount: 6,
+        perspectiveSources: const [
+          SourceMini(name: 'A'),
+          SourceMini(name: 'B'),
+          SourceMini(name: 'C'),
+          SourceMini(name: 'D'),
+          SourceMini(name: 'E'),
+          SourceMini(name: 'F'),
+        ],
+      ),
+    ));
+
+    // Les avatars de la pile + le point d'identité de la source de la carte.
+    expect(find.byType(SourceDot), findsNWidgets(4));
+    expect(find.text('6 sources'), findsOneWidget);
+  });
+
+  testWidgets('footer no longer renders a sub-theme pill', (tester) async {
     await tester.pumpWidget(_wrap(
       FluxContinuArticleCard(article: _content(sourceTheme: 'tech')),
     ));
 
-    expect(find.byKey(const Key('flux-theme-pill')), findsOneWidget);
-    // Le libellé affiché est le thème source mappé (fallback pré-ML), pas le
-    // slug brut : la pastille reste lisible même worker de classif mort.
-    expect(find.text('Tech & Innovation'), findsOneWidget);
+    expect(find.text('Tech & Innovation'), findsNothing);
   });
 
-  testWidgets('no theme pill when neither ML topic nor source theme resolves',
-      (tester) async {
-    await tester.pumpWidget(_wrap(
-      FluxContinuArticleCard(article: _content()),
-    ));
-
-    expect(find.byKey(const Key('flux-theme-pill')), findsNothing);
-  });
 }
